Move queryStringAppender to helpers directory
