+++ src/assessments/readability/sentenceLengthInDescriptionAssessment.js
@@ -1,6 +1,6 @@
 import formatNumber from "../../helpers/formatNumber";
 import { inRangeEndInclusive as inRange } from "../../helpers/inRange";
-import { createAnchorOpeningTag } from "../../queryStringAppender";
+import { createAnchorOpeningTag } from "../../helpers/queryStringAppender";
 import AssessmentResult from "../../values/AssessmentResult";
 import countTooLongSentences from "./../../assessmentHelpers/checkForTooLongSentences";
