+++ src/assessments/readability/transitionWordsAssessment.js
@@ -4,7 +4,7 @@ import formatNumber from "../../helpers/formatNumber";
 import getLanguageAvailability from "../../helpers/getLanguageAvailability";
 import { inRangeStartInclusive as inRange } from "../../helpers/inRange";
 import marker from "../../markers/addMark";
-import { createAnchorOpeningTag } from "../../queryStringAppender";
+import { createAnchorOpeningTag } from "../../helpers/queryStringAppender";
 import { stripIncompleteTags as stripTags } from "../../stringProcessing/stripHTMLTags";
 import AssessmentResult from "../../values/AssessmentResult";
 import Mark from "../../values/Mark";
