+++ src/assessments/seo/KeyphraseDistributionAssessment.js
@@ -1,7 +1,7 @@
 import { merge } from "lodash-es";
 
 import Assessment from "../../assessment";
-import { createAnchorOpeningTag } from "../../queryStringAppender";
+import { createAnchorOpeningTag } from "../../helpers/queryStringAppender";
 import getSentences from "../../stringProcessing/getSentences";
 import AssessmentResult from "../../values/AssessmentResult";
