+++ src/assessments/seo/KeywordDensityAssessment.js
@@ -3,7 +3,7 @@ import { merge } from "lodash-es";
 import Assessment from "../../assessment";
 import formatNumber from "../../helpers/formatNumber";
 import { inRangeEndInclusive, inRangeStartEndInclusive, inRangeStartInclusive } from "../../helpers/inRange";
-import { createAnchorOpeningTag } from "../../queryStringAppender";
+import { createAnchorOpeningTag } from "../../helpers/queryStringAppender";
 import countWords from "../../stringProcessing/countWords";
 import AssessmentResult from "../../values/AssessmentResult";
