+++ src/assessments/seo/outboundLinksAssessment.js
@@ -1,7 +1,7 @@
 import { isEmpty, merge } from "lodash-es";
 
 import Assessment from "../../assessment";
-import { createAnchorOpeningTag } from "../../queryStringAppender";
+import { createAnchorOpeningTag } from "../../helpers/queryStringAppender";
 import AssessmentResult from "../../values/AssessmentResult";
 
 /**
