+++ src/assessments/seo/subheadingsKeywordAssessment.js
@@ -1,7 +1,7 @@
 import { merge } from "lodash-es";
 
 import Assessment from "../../assessment";
-import { createAnchorOpeningTag } from "../../queryStringAppender";
+import { createAnchorOpeningTag } from "../../helpers/queryStringAppender";
 import { getSubheadings } from "../../stringProcessing/getSubheadings";
 import AssessmentResult from "../../values/AssessmentResult";
