+++ src/assessments/seo/urlStopWordsAssessment.js
@@ -1,6 +1,6 @@
 import AssessmentResult from "../../values/AssessmentResult";
 import getLanguageAvailability from "../../helpers/getLanguageAvailability";
-import { createAnchorOpeningTag } from "../../queryStringAppender";
+import { createAnchorOpeningTag } from "../../helpers/queryStringAppender";
 
 const availableLanguages = [ "en" ];
