+++ src/helpers/queryStringAppender/QueryStringAppender.js
