+++ src/helpers/queryStringAppender/singleton.js
